+++ packages/webpack/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @contentmap/webpack
 
+## 0.4.0
+
+### Patch Changes
+
+- Updated dependencies
+  - contentmap@0.4.0
+
 ## 0.3.1
 
 ### Patch Changes